requirements/local.txt
-r ./base.txt

coverage==5.5
click==7.1.2
django-debug-toolbar==3.2
django-debug-toolbar-force==0.1.8
django-nine==0.2.4
dparse==0.5.1
flake8==3.9.0
isort==5.8.0
mccabe==0.6.1
packaging==20.9
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
PyYAML==5.4.1
safety==1.10.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
